requirements.txt
torch>=2.9.1
torchvision>=0.24.1
scikit-learn>=1.8.0
matplotlib>=3.10.8
numpy>=2.4.1
Pillow>=12.1.0
tqdm>=4.67.1
